w3c-reports/12092014/efl/runned-tests/media-source/interfaces-expected.txt

CONSOLE MESSAGE: line 130: ReferenceError: Can't find variable: MediaSource

Harness Error. harness_status.status = 1 , harness_status.message = ReferenceError: Can't find variable: MediaSource
